Refactor parsing to fix v3 / v3.1 detection, and handle unordered metrics
Replaces the regex-based parsing with pattern matching. The new implementation is able to detect:

* Missing mandatory metrics
* Extraneous, unknown metrics
* Invalid metric values
* Invalid metric formats

In all above cases, a `MalformedVectorException` is thrown, containing details about the issue at hand. Previously, trying to parse an invalid vector merely resulted in a `null` return value.

Implementations of `Cvss#getVector` had to be updated to handle cases where temporal and environmental metrics are not present.

Note that this implementation does not enforce metric order for CVSSv2.

Closes #49
Fixes #56