Default optional metrics to "Not Defined"
Score calculation is not designed to deal with `null` values. If a metric is not provided, it is assumed to be "Not Defined" (`ND` for CVSSv2, `X` for CVSSv3). This ensures that score calculation continues to work, when only a subset of optional metrics is provided.

Metrics with "Not Defined" values are omitted when constructing a vector string. This is in line with how visual CVSS calculators operate.

Initialization of metrics with `null` is prevented via `Objects#requireNonNull` checks on setters.

Fixes #65